// src/pages/Whole/Health/data.d.ts
export interface EffectDataType {
  mainScore: number;
  effectNumber: number;
  faultLevel: string;
  subSystemType: string;
  faultValue: string;
  time: Date;
  effect: string;
  faultRate: number;
}
export interface DimensionDataType {
  deminsion: string;
  drivingSafetyScore: number;
  efficiencyScore: string;
  correlationScore: string;
  complexityScore: string;
}